website/requirements.txt
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
imbalanced-learn==0.11.0
mysqlclient==2.2.4
numpy==1.26.0
openpyxl==3.1.2
pandas==2.1.1
plotly==5.18.0
psycopg2-binary==2.9.9
scikit-learn==1.3.2
seaborn==0.13.0
SQLAlchemy==2.0.29
xgboost==2.0.1
WTForms==3.1.1
wordcloud==1.9.3
python-dotenv
